Move more things out of root script and document plugins applied in root
