+++ build.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010 the original author or authors.
+ * Copyright 2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,25 +14,23 @@
  * limitations under the License.
  */
 
-import org.gradle.gradlebuild.UpdateAgpVersions
-import org.gradle.gradlebuild.UpdateBranchStatus
-import org.gradle.gradlebuild.buildquality.incubation.IncubatingApiAggregateReportTask
-import org.gradle.gradlebuild.buildquality.incubation.IncubatingApiReportTask
-
 plugins {
-    base                  // basePluginConvention.distsDirName is used for the location of the distribution, might move this to "subprojects/distributions/build/..."
-    gradlebuild.lifecycle // this needs to be applied first
-    gradlebuild.`ci-reporting`
-    gradlebuild.security
-    gradlebuild.cleanup
-    gradlebuild.buildscan
-    gradlebuild.minify
-    gradlebuild.wrapper
-    gradlebuild.ide
-    gradlebuild.`build-version`
-    gradlebuild.`update-versions`             // Release process: Update 'released-versions.json' to latest Release Snapshots, RCs and Releases
+    base                                     // Build Distribution: 'base.distsDirName' is used for the location of the distribution, might move this to "subprojects/distributions/build/..."
+    gradlebuild.security                     // Check: Ensure sure we only use secure repositories
+
+    gradlebuild.lifecycle                    // CI: Add lifecycle tasks to for the CI pipeline (currently needs to be applied early as it might modify global properties)
     gradlebuild.`generate-subprojects-info`  // CI: Generate subprojects information for the CI testing pipeline fan out
-    gradlebuild.`quick-check`                // Local development: Add `quickCheck` task for a checkstyle etc. only run on all changed files before commit
+    gradlebuild.cleanup                      // CI: Advanced cleanup after the build (like stopping daemons started by tests)
+    gradlebuild.`ci-reporting`               // CI: Prepare reports to be uploaded to TeamCity
+
+    gradlebuild.buildscan                    // Reporting: Add more data through custom tags to build scans
+
+    gradlebuild.`build-version`              // Release process: Set the version for this build
+
+    gradlebuild.ide                          // Local development: Tweak IDEA import
+    gradlebuild.`update-versions`            // Local development: Convenience tasks to update versions in this build: 'released-versions.json', 'agp-versions.properties', ...
+    gradlebuild.wrapper                      // Local development: Convenience tasks to update the wrapper (like 'nightlyWrapper')
+    gradlebuild.`quick-check`                // Local development: Convenience tasks task `quickCheck` for running checkstyle/codenarc only on changed files before commit
 }
 
 buildscript {
@@ -46,9 +44,7 @@ buildscript {
     }
 }
 
-allprojects {
-    group = "org.gradle"
-
+subprojects {
     repositories {
         maven {
             name = "Gradle libs"
@@ -66,55 +62,5 @@ allprojects {
             name = "kotlin-eap"
             url = uri("https://dl.bintray.com/kotlin/kotlin-eap")
         }
-    }
-}
-
-tasks.register("packageBuild") {
-    description = "Build production distros and smoke test them"
-    group =  "build"
-    dependsOn(":distributions:verifyIsProductionBuildEnvironment", ":distributions:buildDists",
-        ":distributions:integTest", ":docs:check", ":docs:checkSamples")
-}
-
-subprojects {
-    plugins.withId("gradlebuild.publish-public-libraries") {
-        tasks.register("promotionBuild") {
-            description = "Build production distros, smoke test them and publish"
-            group = "publishing"
-            dependsOn(":distributions:verifyIsProductionBuildEnvironment", ":distributions:buildDists",
-                ":distributions:integTest", ":docs:check", "publish")
-        }
-    }
-}
-
-tasks.register<UpdateBranchStatus>("updateBranchStatus")
-
-tasks.register<UpdateAgpVersions>("updateAgpVersions") {
-    comment.set(" Generated - Update by running `./gradlew updateAgpVersions`")
-    minimumSupportedMinor.set("3.4")
-    propertiesFile.set(layout.projectDirectory.file("gradle/dependency-management/agp-versions.properties"))
-}
-
-val allIncubationReports = tasks.register<IncubatingApiAggregateReportTask>("allIncubationReports") {
-    val allReports = collectAllIncubationReports()
-    dependsOn(allReports)
-    reports = allReports.associateBy({ it.title.get() }) { it.textReportFile.asFile.get() }
-}
-tasks.register<Zip>("allIncubationReportsZip") {
-    destinationDirectory.set(layout.buildDirectory.dir("reports/incubation"))
-    archiveBaseName.set("incubating-apis")
-    from(allIncubationReports.get().htmlReportFile)
-    from(collectAllIncubationReports().map { it.htmlReportFile })
-}
-
-fun Project.collectAllIncubationReports() = subprojects.flatMap { it.tasks.withType(IncubatingApiReportTask::class) }
-
-// Ensure the archives produced are reproducible
-allprojects {
-    tasks.withType<AbstractArchiveTask>().configureEach {
-        isPreserveFileTimestamps = false
-        isReproducibleFileOrder = true
-        dirMode = Integer.parseInt("0755", 8)
-        fileMode = Integer.parseInt("0644", 8)
     }
 }
