+++ buildSrc/subprojects/packaging/src/main/kotlin/org/gradle/gradlebuild/packaging/MinifyPlugin.kt
@@ -43,36 +43,33 @@ open class MinifyPlugin : Plugin<Project> {
                 "it.unimi.dsi.fastutil.ints.IntSets"
             )
         )
-
-        allprojects {
-            plugins.withId("java-base") {
-                dependencies {
-                    attributesSchema {
-                        attribute(minified)
-                    }
-                    artifactTypes.getByName("jar") {
-                        attributes.attribute(minified, java.lang.Boolean.FALSE)
-                    }
-                    registerTransform(Minify::class) {
-                        from.attribute(minified, false).attribute(artifactType, "jar")
-                        to.attribute(minified, true).attribute(artifactType, "jar")
-                        parameters {
-                            keepClassesByArtifact = keepPatterns
-                        }
+        plugins.withId("java-base") {
+            dependencies {
+                attributesSchema {
+                    attribute(minified)
+                }
+                artifactTypes.getByName("jar") {
+                    attributes.attribute(minified, java.lang.Boolean.FALSE)
+                }
+                registerTransform(Minify::class) {
+                    from.attribute(minified, false).attribute(artifactType, "jar")
+                    to.attribute(minified, true).attribute(artifactType, "jar")
+                    parameters {
+                        keepClassesByArtifact = keepPatterns
                     }
                 }
-                configurations.all {
-                    afterEvaluate {
-                        // everywhere where we resolve, prefer the minified version
-                        if (isCanBeResolved && !isCanBeConsumed) {
+            }
+            configurations.all {
+                afterEvaluate {
+                    // everywhere where we resolve, prefer the minified version
+                    if (isCanBeResolved && !isCanBeConsumed) {
+                        attributes.attribute(minified, true)
+                    }
+                    // local projects are already minified
+                    if (isCanBeConsumed && !isCanBeResolved) {
+                        if (attributes.getAttribute(Category.CATEGORY_ATTRIBUTE)?.name == Category.LIBRARY
+                                && attributes.getAttribute(Bundling.BUNDLING_ATTRIBUTE)?.name == Bundling.EXTERNAL) {
                             attributes.attribute(minified, true)
-                        }
-                        // local projects are already minified
-                        if (isCanBeConsumed && !isCanBeResolved) {
-                            if (attributes.getAttribute(Category.CATEGORY_ATTRIBUTE)?.name == Category.LIBRARY
-                                    && attributes.getAttribute(Bundling.BUNDLING_ATTRIBUTE)?.name == Bundling.EXTERNAL) {
-                                attributes.attribute(minified, true)
-                            }
                         }
                     }
                 }
