+++ buildSrc/subprojects/plugins/src/main/kotlin/org/gradle/gradlebuild/versioning/BuildVersionPlugin.kt
@@ -82,6 +82,7 @@ fun Project.setBuildVersion() {
     }
 
     project.allprojects {
+        group = "org.gradle"
         version = versionNumber
     }
