+++ buildSrc/subprojects/plugins/src/test/groovy/org/gradle/plugins/performance/PerformanceTestIntegrationTest.groovy
@@ -5,7 +5,6 @@ class PerformanceTestIntegrationTest extends AbstractIntegrationTest {
         buildFile << """
             plugins {
                 id 'java-library'
-                id 'gradlebuild.int-test-image'
                 id 'gradlebuild.performance-test'
             }
             subprojects {
